data_crawling/05/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
psycopg2==2.8.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
scrapy==1.7.3
python-dotenv==0.10.3
